convert token
